requirements/dev.txt
Django==1.7.3
MySQL-python==1.2.5
argparse==1.2.1
distribute==0.6.24
django-model-utils==2.2
django-nose==1.3
factory-boy==2.4.1
ipython==2.3.1
nose==1.3.4
wsgiref==0.1.2
